EPI-160:II: Move composite machinery out of fhir context
