+++ packages/shared-src/src/services/database.js
@@ -10,10 +10,11 @@ import { migrate, assertUpToDate } from './migrations';
 import * as models from '../models';
 import { initSyncHooks } from '../models/sync';
 import { createDateTypes } from './createDateTypes';
-import { initFhirTypes, setupQuote } from './fhirTypes';
+import { createFhirTypes } from './fhirTypes';
+import { setupQuote } from '../utils/pgComposite';
 
 createDateTypes();
-initFhirTypes();
+createFhirTypes();
 
 // this allows us to use transaction callbacks without manually managing a transaction handle
 // https://sequelize.org/master/manual/transactions.html#automatically-pass-transactions-to-all-queries
