+++ packages/shared-src/src/utils/pgComposite/parser.js
@@ -1,0 +1,3 @@
+export function parseComposite() {
+  // TODO
+}
