+++ packages/shared-src/src/utils/pgComposite/stringifier.js
@@ -1,18 +1,6 @@
-import { snakeCase } from 'lodash';
-import { DataTypes, Utils } from 'sequelize';
+import { Utils } from 'sequelize';
 import { formatISO9075 } from 'date-fns';
-import * as yup from 'yup';
-
-/**
- * Register a new type with sequelize.
- * CANNOT register an extension of an existing sequelize type (like STRING, etc).
- * Only use with fully-new types.
- */
-export function newPgDataType(Klass) {
-  const name = Klass.key;
-  DataTypes[name] = Utils.classToInvokable(Klass);
-  DataTypes[name].types.postgres = [Klass.pgName];
-}
+import { Composite } from './sequelizeType';
 
 let QUERY_GENERATOR;
 /**
@@ -101,74 +89,4 @@ function compositeString(string) {
 
   // Let's do this simply by double-quoting everything.
   return `"${string.replace('"', '\\"').replace('\\', '\\\\')}"`;
-}
-
-const ABSTRACT = DataTypes.ABSTRACT.prototype.constructor;
-export class COMPOSITE extends ABSTRACT {
-  static ValueClass;
-
-  static get key() {
-    return this.name;
-  }
-
-  static get pgName() {
-    return `fhir.${snakeCase(this.ValueClass.name.replace(/^Fhir/, ''))}`;
-  }
-
-  constructor(options) {
-    super();
-    this.options = options || {};
-    this.key = this.constructor.key;
-  }
-
-  toSql() {
-    return this.constructor.pgName;
-  }
-
-  validate(value) {
-    if (value instanceof this.constructor.ValueClass) return true;
-    throw new ValidationError(`value is not a ${this.constructor.ValueClass}`);
-  }
-
-  fieldSet(value) {
-    if (value instanceof this.constructor.ValueClass) {
-      return value.sqlFields();
-    } else {
-      return new this.constructor.ValueClass(value).sqlFields();
-    }
-  }
-
-  _stringify(value, options) {
-    return compositeToSql(this.fieldSet(value), options);
-  }
-
-  _value(value, options) {
-    return compositeToSql(this.fieldSet(value), options);
-  }
-
-  _sanitize(value, options) {
-    return this.constructor.ValueClass.fromSql(value, options);
-  }
-}
-
-export class Composite {
-  static SCHEMA = yup.object();
-  static FIELD_ORDER = [];
-
-  constructor(params) {
-    this.params = this.constructor.SCHEMA.validateSync(params);
-  }
-
-  sqlFields() {
-    return this.constructor.FIELD_ORDER.map(name => this.params[name]);
-  }
-
-  static fromSql(value) {
-    console.error({ fromSql: value });
-    return value;
-  }
-
-  static fake() {
-    throw new Error('Must be overridden');
-  }
 }
